==> Notebooks/1_Transformaciones.ipynb <==
Importación de librerias

import pandas as pd
import gzip
import json
import ast

## Transformación **steam_games**

### Carga de datos de juegos de Steam desde un archivo JSON comprimido en gzip

# Creación de DataFrame a partir de un Archivo JSON Comprimido
filas = []

with gzip.open("datasets/steam_games.json.gz", 'rt', encoding='utf-8') as archivo:
    for linea in archivo:
        filas.append(json.loads(linea))
        
# Crear el DataFrame
df_games = pd.DataFrame(filas)

df_games.head()

### Verificación y Limpieza del DataFrame

# Verificamos las dimensiones de nuestro DF
df_games.shape

# Eliminamos las filas que contienen SOLO VALORES NULOS
df_games = df_games.dropna(how="all").reset_index(drop=True) 
df_games.head()

# Verificamos nuevamente las dimensiones de nuestro DF
df_games.shape

Después de eliminar las filas que contenían únicamente valores nulos, las dimensiones del DataFrame se redujeron de (120445 filas, 13 columnas) a (32135 filas, 13 columnas). Este paso de limpieza ha reducido el tamaño del conjunto de datos y ha eliminado las filas que no contenían información útil en ninguna de las columnas.

# Análisis de Calidad de Datos en un DataFrame"
def verificar_datos(df):
    verif = {"nombre_campo": [], "tipo_dato": [], "%_No_Nulos": [], "%_Nulos": [], "Nulos": []}

    for columna in df.columns:
        porcentaje_no_nulos = (df[columna].count() / len(df)) * 100
        verif["nombre_campo"].append(columna)
        verif["tipo_dato"].append(df[columna].dtypes)
        verif["%_No_Nulos"].append(round(porcentaje_no_nulos, 2))
        verif["%_Nulos"].append(round(100 - porcentaje_no_nulos, 2))
        verif["Nulos"].append(df[columna].isnull().sum())

    df_info = pd.DataFrame(verif)
        
    return df_info

verificar_datos(df_games)

### Eliminación de columnas
Se eliminaran las columnas **'title'**, **'url'**, **'early_access'**, **'reviews_url'** y **'specs'** del DataFrame df_games. La decisión de eliminar 'title' se basa en que hay otra columna similar con menos datos nulos, y las otras columnas no se requerirán para los endpoints que se ha pedido. Este paso de eliminación se realizó para simplificar el conjunto de datos y trabajar con las columnas relevantes.

df_games.drop(['title', 'url', 'early_access', 'reviews_url','specs'], axis=1, inplace=True)

df_games.head()

### Ajustes para poder guardar el DF en formato Parquet y CSV

La columna **'price'** del DataFrame **df_games** se convirtió al tipo de dato STR para facilitar su manejo. Luego, el DataFrame se guardó en los formatos Parquet y CSV.

# Convertir la columna 'price' a tipo string
df_games['price'] = df_games['price'].astype(str)
df_games.to_parquet('datasets/data_transf/t_df_games.parquet', index=False)
df_games.to_csv('datasets/data_transf/t_df_games.csv', index=False)

## Transformación **user_reviews**

#### Carga de datos de 'reviews' de Steam desde un archivo JSON comprimido en gzip

# Creación de DataFrame a partir de un Archivo JSON Comprimido
row = []

with gzip.open("datasets/user_reviews.json.gz", 'rt', encoding='utf-8') as archivo:
    for linea in archivo:
        # ast.literal_eval() convierte cada línea, que es una cadena de texto, en un objeto de Python
        row.append(ast.literal_eval(linea))

# Crear el DataFrame
df_reviews = pd.DataFrame(row)

df_reviews.head()

# Normalización de datos JSON en un DataFrame
df_reviews = pd.json_normalize(row, record_path='reviews', meta=['user_id','user_url'])
df_reviews.head()

#### Verificación y Limpieza del DataFrame

# Verificamos las dimensiones de nuestro DF
df_reviews.shape

# Eliminación de columnas que son innecesarias para los siguientes procesos
df_reviews.drop(['funny','last_edited','helpful','user_url'], axis=1, inplace=True)

# Eliminación de Filas Nulas y Duplicadas
df_reviews = df_reviews.dropna(how='all')
df_reviews = df_reviews.drop_duplicates()
df_reviews.head()

# Verificamos nuevamente las dimensiones de nuestro DF
df_reviews.shape

Después de eliminar las columnas innecesarias y filas que contenían únicamente valores nulos y duplicados, las dimensiones del DataFrame se redujeron de (59305 filas, 9 columnas) a (58431 filas, 5 columnas). Este paso de limpieza ha reducido el tamaño del conjunto de datos.

verificar_datos(df_reviews)

#### Guardamos los dataframes transformados en archivos Parquet y CSV

df_reviews.to_parquet('datasets/data_transf/t_user_reviews.parquet', index=False)
df_reviews.to_csv('datasets/data_transf/t_user_reviews.csv', index=False)

## Transformación **users_items**

# Creación de DataFrame a partir de un Archivo JSON Comprimido
row = []

with gzip.open("datasets/users_items.json.gz", 'rt', encoding='utf-8') as archivo:
    for linea in archivo:
        # ast.literal_eval() convierte cada línea, que es una cadena de texto, en un objeto de Python
        row.append(ast.literal_eval(linea))

# Crear el DataFrame
df_items = pd.DataFrame(row)

df_items.head()

# Normalización de datos JSON en un DataFrame
df_items = pd.json_normalize(row, record_path='items', meta=['user_id','items_count','steam_id','user_url'])
df_items.head()

#### Verificación y Limpieza del DataFrame

# Verificamos las dimensiones de nuestro DF
df_items.shape

# Eliminación de columnas que son innecesarias para los siguientes procesos
df_items.drop(['user_url', 'item_name', 'steam_id', 'playtime_2weeks'], axis=1, inplace=True)

# Eliminación de Filas Nulas y Duplicadas
df_items = df_items.dropna(how='all')
df_items = df_items.drop_duplicates()
df_items.head()

# Verificamos nuevamente las dimensiones de nuestro DF
df_items.shape

Después de eliminar las columnas innecesarias y filas que contenían únicamente valores nulos y duplicados, las dimensiones del DataFrame se redujeron de (5153209 filas, 8 columnas) a (5094092 filas, 4 columnas). Este paso de limpieza ha reducido el tamaño del conjunto de datos.

verificar_datos(df_items)

#### Guardamos los dataframes transformados en formato Parquet y CSV

df_items.to_parquet('datasets/data_transf/t_users_items.parquet', index=False)
df_items.to_csv('datasets/data_transf/t_users_items.csv', index=False)
