==> Notebooks/2_Feature_Engineering.ipynb <==
### Importación de librerias necesarias

import pandas as pd
from textblob import TextBlob
import nltk
from nltk.sentiment import SentimentIntensityAnalyzer

## Análisis de sentimiento con libreria **textblob**

Carga y lectura del DF

df_review1 = pd.read_csv('datasets/data_transf/t_user_reviews.csv')
df_review1

# Función de Análisis de Sentimiento: Determinando Polaridad en Texto

def analisis_sentimiento(review):
    analisis = TextBlob(str(review))
    # Asigna un valor según el análisis de sentimiento
    if analisis.sentiment.polarity > 0.1: 
        return 2  # Positivo
    elif analisis.sentiment.polarity < -0.1:
        return 0  # Malo
    else:
        return 1  # Neutral

# Análisis de Sentimiento aplicado a columna 'review': Creación de una Columna 'Sentiment_analysis'
df_review1['Sentiment_analysis'] = df_review1['review'].apply(analisis_sentimiento)

df_review1.head()

df_review1['Sentiment_analysis'].value_counts()

## Análisis de sentimiento con biblioteca **nltk**

Carga y lectura del DF

df_review2 = pd.read_csv('datasets/data_transf/t_user_reviews.csv')
df_review2

# Análisis de Sentimiento con VADER en NLTK

from nltk.sentiment import SentimentIntensityAnalyzer

def analyze_sentiment_vader(review):
    sid = SentimentIntensityAnalyzer()
    scores = sid.polarity_scores(str(review))

    compound_score = scores['compound']

    if compound_score > 0.1:
        return 2  # Positivo
    elif compound_score < -0.1:
        return 0  # Negativo
    else:
        return 1  # Neutral


# Aplicación de Análisis de Sentimiento VADER a 'review'
df_review2['Sentiment_analysis'] = df_review2['review'].apply(analyze_sentiment_vader)

df_review2.head()

df_review2['Sentiment_analysis'].value_counts()

# Eliminación de Columna 'review'. La nueva columna Análisis de Sentimiento VADER tomara el lugar de columna 'review'
df_review2.drop(['review'], axis=1, inplace=True)

df_review2.head()

#### Guardamos los dataframes transformados en formato Parquet y CSV

df_review2.to_parquet('datasets/f_e_user_reviews.parquet', index=False)
df_review2.to_csv('datasets/f_e_user_reviews.csv', index=False)
