==> Notebooks/4_Optimizacion_df_ml.ipynb <==
import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.preprocessing import StandardScaler

# DF original para el proceso Merge
df_original = pd.read_csv('datasets/data_fusionada.csv')

## Filtrar para el proceso de recomendación **item-item**

### Trabajando columna **'recommend'**

# Cargar el DataFrame
columnas = ['item_id', 'genres', 'playtime_forever', 'recommend']
df_item = pd.read_csv('datasets/data_fusionada.csv', usecols=columnas)

df_item.head()

# Vemos las dimensiones del DF
df_item.shape

# Agrupamos por interacciones 'item_id' y contamos las 'recommed' para cada juego
frecuencia_juegos = df_item[df_item['recommend'] == True]['item_id'].value_counts().reset_index()
frecuencia_juegos.columns = ['item_id', 'frecuencia_rec']
frecuencia_juegos

# Unimos el dataframe principal con el DF 'frecuencia de juegos'
df_merge1 = pd.merge(df_item, frecuencia_juegos, on='item_id', how='left')
df_merge1['frecuencia_rec'] = df_merge1['frecuencia_rec'].fillna(0)
df_merge1

# Convertir la columna 'frecuencia_rec' a tipo de dato entero.
df_merge1['frecuencia_rec'] = df_merge1['frecuencia_rec'].astype(int)
df_merge1.head()

df_merge1['recommend'].value_counts()

Filtrar datos relevantes en df_merge1 basándose en la frecuencia de recomendación

filtro_rec = df_merge1[df_merge1['frecuencia_rec'] >= 5]
filtro_rec

# Verificamos la cantidad de 'item_id' tenemos para trabajar
filtro_rec['item_id'].nunique()

# Seleccionar columnas relevantes en df_f_rec para análisis posterior
columnas = ['item_id', 'genres', 'frecuencia_rec']
df_f_rec = filtro_rec[columnas]

# Eliminar duplicados manteniendo la primera ocurrencia
df_f_rec_agrupado = df_f_rec.drop_duplicates(subset=['item_id'])
df_f_rec_agrupado

df_f_rec_agrupado.shape

Terminamos el trabajo de la columna **recommend**, consiguiendo disminuir a 802 juegos **'item_id'**, este filtro nos facilitará el proceso de ML

### Trabajo columna **'playtime_forever'**

# Cargar el DataFrame
columnas = ['item_id', 'genres', 'playtime_forever', 'recommend']
df_item2 = pd.read_csv('data_fusionada.csv', usecols=columnas)
df_item2.head()

# Sumar el tiempo de juego para cada juego (item_id)
playtime = df_item2.groupby('item_id')['playtime_forever'].sum().reset_index()
playtime.columns = ['item_id', 'total_playtime']
playtime


# Consultando la cantidad de juegos tenemos 'item_id'
playtime['item_id'].nunique()

# Ordenar por la columna 'playtime_forever' de manera ascendente
playtime_ord = playtime.sort_values(by='total_playtime')
playtime_ord.head(10)

# Al verificar que nuestra columna total_playtime tiene datos booleanos, los cambiamos a enteros, para optimizar los futuros procesos.
playtime_ord['total_playtime'] = playtime_ord['total_playtime'].astype(int)

#Filtrar juegos con tiempo de juego superior a 2000 minutos
filtro_time = playtime_ord[playtime_ord['total_playtime'] >= 2000]
filtro_time

Luego de trabajar con **'playtime_forever**  hemos logrado identificar a los datos de juegos más optimos para los siguientes procesos

# Crear un DataFrame con las columnas necesarias
columnas = ['item_id', 'genres']
df_genres = df_original[columnas]

frecuencia_genres = df_genres.groupby(['item_id', 'genres']).size().reset_index(name='frecuencia_genres')

# Consolidar la información agrupando por item_id
df_genres_agrupado = frecuencia_genres.groupby('item_id', as_index=False).agg({'genres': ' | '.join, 'frecuencia_genres': 'sum'})
df_genres_agrupado

# Verificamos la cantidad de juegos que entrarían a futuros proceso ML
df_genres_agrupado['item_id'].nunique()

filtro_genres= df_genres_agrupado[df_genres_agrupado['frecuencia_genres'] >= 5]
filtro_genres

# Verificamos nuevamente como ha cambiado la cantidad de juegos
filtro_genres['item_id'].nunique()

Luego de trabajar con **'frecuencia_genres'**  hemos logrado identificar 880 juegos con mejor data, para nuestros siguientes proceso

## Guardar DF que servirá para realizar el proceso de ML recomendación 'item'-'item'

# Unir los DataFrames
df_final = pd.merge(df_f_rec_agrupado, filtro_time, on='item_id')
df_final = pd.merge(df_final, filtro_genres, on='item_id')
df_final

# Supongamos que quieres renombrar las columnas 'genres_x' y 'genres_y'
df_final = df_final.rename(columns={'frecuencia_genres': 'genres', 'frecuencia_rec': 'recommend','total_playtime':'playtime_forever'})
df_final

columnas = ['item_id', 'genres', 'recommend', 'playtime_forever']
df_final = df_final[columnas]
df_final

# Mergeamos con la data original, usando 'item_id'. Gracias al trabajo realizado anteriormente, vamos a obtener filas de nuestros juegos mejor data.
df_resultado = pd.merge(df_final, df_original[['item_id', 'app_name']], on='item_id', how='inner')
df_resultado

# Eliminamos duplicados y nulos
df_resultado = df_resultado.drop_duplicates()
df_resultado = df_resultado.dropna()
df_resultado

# Ordenamos las columnas, para facilitar el futuro proceso de ML
columnas = ['item_id','app_name','recommend','playtime_forever','genres']
df_ok_item_item = df_resultado[columnas]
df_ok_item_item

# Guardamos nuestro df para el proceso de ML recomendaciones
df_ok_item_item.to_csv('dataML_item_item.csv', index=False)

# Filtrar para el proceso de recomendación **usuario-item**

# Cargamos DF
df_us = pd.read_csv('datasets/data_fusionada.csv')

df_us

### Trabajamos la columna **'recommend'**

usua_fil = df_us[df_us['recommend'] == True]['user_id'].value_counts().reset_index()
usua_fil.columns = ['user_id', 'frecuencia_rec']
usua_fil

# Aseguramos que la columna tenga datos INT
usua_fil['frecuencia_rec'] = usua_fil['frecuencia_rec'].astype(int)
#Filtrar usuarios con más de 2 interacciones
filtro_rec = usua_fil[usua_fil['frecuencia_rec'] >= 3]
filtro_rec

# Sumar el tiempo de juego para cada juego (user_id)
us_time = df_us.groupby('user_id')['playtime_forever'].sum().reset_index()
us_time.columns = ['user_id', 'total_playtime']
us_time

# Al verificar que nuestra columna total_playtime tiene datos booleanos, los cambiamos a enteros, para optimizar los futuros procesos.
us_time['total_playtime'] = us_time['total_playtime'].astype(int)
#Filtrar juegos con tiempo de juego superior a 4000 minutos
us_time_fil = us_time[us_time['total_playtime'] >= 4000]
us_time_fil

df_merged_us = pd.merge(filtro_rec, us_time_fil, on='user_id')
df_merged_us

# Eliminamos duplicados y nulos
df_merged_us = df_merged_us.drop_duplicates()
df_merged_us = df_merged_us.dropna()
df_merged_us

# Mergeamos nuestra data original
df_finalx = pd.merge(df_merged_us, df_original, on='user_id')
df_finalx

# Selección de las columnas para el siguiente proceso de ML
columnas = ['app_name', 'item_id', 'playtime_forever', 'user_id','recommend', 'Sentiment_analysis']
df_final_users = df_finalx[columnas]
df_final_users

# Guardamos nuestro df para el proceso de ML recomendaciones
df_final_users.to_csv('dataML_user_item.csv', index=False)
